+++ src/account/index.ts
@@ -1,46 +1,45 @@
-import {
-	generateBlock,
-	generateBlockHeader,
-	generateMetadata,
-	generateSeedBlock,
-	generateSignatureKeyPair,
-	signBlockHeader,
-	signSeedBlock,
-	verifyBlock,
-	verifySeedBlockSignature,
-} from '../core'
 import { IAccount } from '../models/IAccount'
 import { IBlock } from '../models/IBlock'
+import { Core } from '../core'
 import { IKeyPair } from '../models/IKeyPair'
 import { ISeedBlock } from '../models/ISeedBlock'
+import { IDnsRecord } from '../models/IDnsRecord'
+import { IDnsRecordMessage } from '../models/IDnsRecordMessage'
+import { IDnsProvider } from '../models/IDnsProvider'
 
 export class Account {
+	private core: Core
+
 	public accountPublicKey?: string
 	public blockPublicKey?: string
 	private blockPrivateKey?: string
-	// private accountProviders?: []
-	// private accountDomains?: []
+
+	// public accountDomains: IDomainName[] = []
+	// public accountDomainRoutes: IDomainNameRoute[] = []
+
 	private seedBlock?: ISeedBlock
 	private accountBlocks: IBlock[] = []
 
-	constructor() {}
+	constructor(core: Core | undefined = undefined) {
+		this.core = core ?? new Core()
+	}
 
 	async init(
 		saveBlockPrivateKey?: boolean,
 	): Promise<{ accountKeyPair: IKeyPair; blockKeyPair: IKeyPair }> {
-		const accountKeyPair = await generateSignatureKeyPair()
-		const blockKeyPair = await generateSignatureKeyPair()
+		const accountKeyPair = await Core.generateSignatureKeyPair()
+		const blockKeyPair = await Core.generateSignatureKeyPair()
 		this.accountPublicKey = accountKeyPair.publicKey
 		this.blockPublicKey = blockKeyPair.publicKey
 		this.accountBlocks = []
-		this.seedBlock = await generateSeedBlock(
+		this.seedBlock = await Core.generateSeedBlock(
 			accountKeyPair.publicKey,
 			blockKeyPair.publicKey,
 			1,
 			1,
 		)
 
-		this.seedBlock = await signSeedBlock(
+		this.seedBlock = await Core.signSeedBlock(
 			this.seedBlock,
 			accountKeyPair.privateKey,
 		)
@@ -80,7 +79,7 @@ export class Account {
 
 		const nextBlockId = this.accountBlocks.length + 1
 
-		const unsignedblockHeader = await generateBlockHeader(
+		const unsignedblockHeader = await Core.generateBlockHeader(
 			content,
 			this.accountPublicKey!,
 			this.blockPublicKey!,
@@ -88,13 +87,13 @@ export class Account {
 			1,
 			1,
 		)
-		const signedBlockHeader = await signBlockHeader(
+		const signedBlockHeader = await Core.signBlockHeader(
 			unsignedblockHeader,
 			blockPrivateKey ?? this.blockPrivateKey!,
 		)
 
-		const blockMetadata = generateMetadata(this.seedBlock!)
-		const block = generateBlock(signedBlockHeader, content, blockMetadata)
+		const blockMetadata = Core.generateBlockMetadata(this.seedBlock!)
+		const block = Core.generateBlock(signedBlockHeader, content, blockMetadata)
 		this.accountBlocks.push(block)
 		return block
 	}
@@ -119,7 +118,7 @@ export class Account {
 
 		const olderBlock = this.accountBlocks[blockIndex]
 
-		const unsignedblockHeader = await generateBlockHeader(
+		const unsignedblockHeader = await Core.generateBlockHeader(
 			content,
 			this.accountPublicKey!,
 			this.blockPublicKey!,
@@ -127,12 +126,12 @@ export class Account {
 			olderBlock.header.update_id + 1,
 			1,
 		)
-		const signedBlockHeader = await signBlockHeader(
+		const signedBlockHeader = await Core.signBlockHeader(
 			unsignedblockHeader,
 			blockPrivateKey ?? this.blockPrivateKey!,
 		)
-		const blockMetadata = generateMetadata(this.seedBlock!)
-		const block = generateBlock(signedBlockHeader, content, blockMetadata)
+		const blockMetadata = Core.generateBlockMetadata(this.seedBlock!)
+		const block = Core.generateBlock(signedBlockHeader, content, blockMetadata)
 		this.accountBlocks[blockIndex] = block
 		return block
 	}
@@ -153,7 +152,7 @@ export class Account {
 
 		const olderBlock = this.accountBlocks[blockIndex]
 
-		const unsignedblockHeader = await generateBlockHeader(
+		const unsignedblockHeader = await Core.generateBlockHeader(
 			'',
 			this.accountPublicKey!,
 			this.blockPublicKey!,
@@ -161,12 +160,12 @@ export class Account {
 			olderBlock.header.update_id + 1,
 			1,
 		)
-		const signedBlockHeader = await signBlockHeader(
+		const signedBlockHeader = await Core.signBlockHeader(
 			unsignedblockHeader,
 			blockPrivateKey ?? this.blockPrivateKey!,
 		)
-		const blockMetadata = generateMetadata(this.seedBlock!)
-		const block = generateBlock(signedBlockHeader, '', blockMetadata)
+		const blockMetadata = Core.generateBlockMetadata(this.seedBlock!)
+		const block = Core.generateBlock(signedBlockHeader, '', blockMetadata)
 		this.accountBlocks[blockIndex] = block
 		return block
 	}
@@ -181,14 +180,14 @@ export class Account {
 
 		const block = this.accountBlocks[blockIndex]
 
-		return await verifyBlock(block)
+		return await Core.verifyBlock(block)
 	}
 
 	async verifySeedBlock(): Promise<boolean> {
 		if (this.seedBlock === undefined) {
 			throw new Error('Seed block undefined')
 		}
-		return await verifySeedBlockSignature(this.seedBlock!)
+		return await Core.verifySeedBlockSignature(this.seedBlock!)
 	}
 
 	async exportAccount(): Promise<IAccount> {
@@ -204,5 +203,139 @@ export class Account {
 	): Promise<void> {
 		this.seedBlock = seedBlock
 		this.accountBlocks = accountBlocks
+	}
+
+	public static async getNameServerByDomain(
+		domainName: string,
+	): Promise<IDnsProvider | undefined> {
+		const nameServer = await Core.getNameServerByDomain(domainName)
+		return nameServer
+	}
+
+	public async getDnsRecord(
+		domainName: string,
+		nameServerAddress: string,
+	): Promise<IDnsRecord | undefined> {
+		return await this.core.getDnsRecord(domainName, nameServerAddress)
+	}
+
+	public async createDnsRecord(
+		domainName: string,
+		hostingProviderAddresses: string[],
+		blockPrivateKey?: string,
+	): Promise<void> {
+		if (!this.isAccountInitialized()) {
+			throw new Error('Account not initialized')
+		}
+		if (blockPrivateKey === undefined && !this.isBlockPrivateKeyInitialized()) {
+			throw new Error('Private key not initialized')
+		}
+
+		const dnsRecord: IDnsRecord = {
+			domainName: domainName,
+			accountPublicKey: this.accountPublicKey!,
+			hostingProviderAddresses: hostingProviderAddresses,
+		}
+		const messageMetadata = Core.generateMessageMetadata(this.seedBlock!)
+		const unsignedMessageHeader = await Core.generateMessageHeader(
+			dnsRecord,
+			this.accountPublicKey!,
+			blockPrivateKey ?? this.blockPrivateKey!,
+		)
+		const signedMessageHeader = await Core.signMessageHeader(
+			unsignedMessageHeader,
+			blockPrivateKey ?? this.blockPrivateKey!,
+		)
+
+		const dnsRecordMessage: IDnsRecordMessage = Core.generateMessage(
+			signedMessageHeader,
+			dnsRecord,
+			messageMetadata,
+		)
+		const nameServer = await Account.getNameServerByDomain(domainName)
+
+		return await this.core.createDnsRecord(
+			dnsRecordMessage,
+			nameServer!.nameServerAddress[0],
+		)
+	}
+
+	public async updateDnsRecord(
+		domainName: string,
+		hostingProviderAddresses: string[],
+		blockPrivateKey?: string,
+	): Promise<void> {
+		if (!this.isAccountInitialized()) {
+			throw new Error('Account not initialized')
+		}
+		if (blockPrivateKey === undefined && !this.isBlockPrivateKeyInitialized()) {
+			throw new Error('Private key not initialized')
+		}
+		const dnsRecord: IDnsRecord = {
+			domainName: domainName,
+			accountPublicKey: this.accountPublicKey!,
+			hostingProviderAddresses: hostingProviderAddresses,
+		}
+		const messageMetadata = Core.generateMessageMetadata(this.seedBlock!)
+		const unsignedMessageHeader = await Core.generateMessageHeader(
+			dnsRecord,
+			this.accountPublicKey!,
+			blockPrivateKey ?? this.blockPrivateKey!,
+		)
+		const signedMessageHeader = await Core.signMessageHeader(
+			unsignedMessageHeader,
+			blockPrivateKey ?? this.blockPrivateKey!,
+		)
+
+		const dnsRecordMessage: IDnsRecordMessage = Core.generateMessage(
+			signedMessageHeader,
+			dnsRecord,
+			messageMetadata,
+		)
+		const nameServer = await Account.getNameServerByDomain(domainName)
+
+		return await this.core.updateDnsRecord(
+			dnsRecordMessage,
+			nameServer!.nameServerAddress[0],
+		)
+	}
+
+	public async deleteDnsRecord(
+		domainName: string,
+		hostingProviderAddresses: string[],
+		blockPrivateKey?: string,
+	): Promise<void> {
+		if (!this.isAccountInitialized()) {
+			throw new Error('Account not initialized')
+		}
+		if (blockPrivateKey === undefined && !this.isBlockPrivateKeyInitialized()) {
+			throw new Error('Private key not initialized')
+		}
+		const dnsRecord: IDnsRecord = {
+			domainName: domainName,
+			accountPublicKey: this.accountPublicKey!,
+			hostingProviderAddresses: hostingProviderAddresses,
+		}
+		const messageMetadata = Core.generateMessageMetadata(this.seedBlock!)
+		const unsignedMessageHeader = await Core.generateMessageHeader(
+			dnsRecord,
+			this.accountPublicKey!,
+			blockPrivateKey ?? this.blockPrivateKey!,
+		)
+		const signedMessageHeader = await Core.signMessageHeader(
+			unsignedMessageHeader,
+			blockPrivateKey ?? this.blockPrivateKey!,
+		)
+
+		const dnsRecordMessage: IDnsRecordMessage = Core.generateMessage(
+			signedMessageHeader,
+			dnsRecord,
+			messageMetadata,
+		)
+		const nameServer = await Account.getNameServerByDomain(domainName)
+		return await this.core.deleteDnsRecord(
+			dnsRecordMessage,
+			nameServer!.nameServerAddress[0],
+		)
 	}
 }