+++ src/constants/dns.ts
@@ -1,0 +1,8 @@
+import { IDnsProvider } from '../models/IDnsProvider'
+
+export const NAME_SERVERS: IDnsProvider[] = [
+	{
+		domainExtension: 'stw',
+		nameServerAddress: ['localhost:3000', 'dns.spacetimewave.com'],
+	},
+]