+++ src/models/IDomainNameEntry.ts
@@ -1,5 +1,0 @@
-export interface IDomainNameEntry {
-	domainName: string
-	domainUrls: string[]
-	domainIPs: string[]
-}