+++ src/models/IMessage.ts
@@ -1,0 +1,8 @@
+import { IMessageHeader } from './IMessageHeader'
+import { IMessageMetadata } from './IMessageMetadata'
+
+export interface IMessage {
+	header: IMessageHeader
+	content: any
+	metadata: IMessageMetadata
+}