+++ src/models/IMessageHeader.ts
@@ -1,0 +1,7 @@
+export interface IMessageHeader {
+	version: number
+	output_hash: string
+	address: string
+	public_key: string
+	signature: string | undefined
+}