+++ src/models/IMessageMetadata.ts
@@ -1,0 +1,5 @@
+import type { ISeedBlock } from './ISeedBlock'
+
+export interface IMessageMetadata {
+	seed_block: ISeedBlock
+}